Report writing - Rough/Codes/Codes and schematics/Website files/flipio/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
dj-database-url==0.4.2
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.6.4
gunicorn==19.7.1
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
psycopg2==2.7.3.1
pytz==2017.2
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
whitenoise==3.3.1
zope.interface==4.4.3
